Refactored and added Pester test
Added a Pester test for the Remove-CT365Groupp function

Implemented parameter validation to ensure proper input values.
Changed the deletion of the groups based on their types to reduce code duplication and enhance maintainability.
Removed the check for group existence as we can't delete a group if it not exists.
Reduced the number of Write-PSFMessage calls in the function to enhance code readability and maintainability.